Move all non `.spec.ts` files into `tests/util`
