+++ frontend/tests/login.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from "@playwright/test";
-import { navigateTo, test } from "./common";
+import { navigateTo, test } from "./util/common";
 
 test("Login", async ({ page, baseURL, browserName }) => {
     test.skip(browserName === "webkit", "Skip safari because it doesn't allow http logins");
